agents/video_producer_agent/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
Authlib==1.5.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
decorator==5.2.1
Deprecated==1.2.18
docstring_parser==0.16
fastapi==0.115.12
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
google-adk==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-trace==1.16.1
google-cloud-video-transcoder==1.15.1
google-crc32c==1.7.1
google-genai==1.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.29.1
gradio_client==1.10.1
graphviz==0.20.3
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.7.1
mdurl==0.1.2
moviepy==2.1.2
mutagen==1.47.0
numpy==2.2.5
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==10.4.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.10
safehttpx==0.1.6
semantic-version==2.10.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
tinytag==2.1.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
